+++ conda/validate.go
@@ -5,14 +5,14 @@ import (
 
 	"github.com/robocorp/rcc/common"
 	"github.com/robocorp/rcc/pretty"
+	"github.com/robocorp/rcc/settings"
 )
 
-const (
-	longPathSupportArticle = `https://robocorp.com/docs/product-manuals/robocorp-lab/troubleshooting#windows-has-to-have-long-filenames-support-on`
-)
+const ()
 
 var (
-	validPathCharacters = regexp.MustCompile("(?i)^[.a-z0-9_:/\\\\~-]+$")
+	longPathSupportArticle = settings.Global.DocsLink("product-manuals/robocorp-lab/troubleshooting#windows-has-to-have-long-filenames-support-on")
+	validPathCharacters    = regexp.MustCompile("(?i)^[.a-z0-9_:/\\\\~-]+$")
 )
 
 func ValidLocation(value string) bool {